fix : user navigate logic
