+++ src/pages/LoginPage.tsx
@@ -4,10 +4,8 @@ import { LoginForm } from '@/features/auth/components/LoginForm';
 import { useLoginForm } from '@/features/auth/hooks/useLoginForm';
 
 export const LoginPage = () => {
-  const { form, onSubmit, checkIfUserIsLoggedIn } = useLoginForm();
-  useEffect(() => {
-    checkIfUserIsLoggedIn();
-  }, [checkIfUserIsLoggedIn]);
+  const { form, onSubmit } = useLoginForm();
+
   return (
     <div className="min-h-screen flex items-center justify-center bg-background">
       <div className="max-w-md w-full p-6 bg-background rounded-md shadow-lg">
